requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
chromadb==1.0.20
pydantic==2.8.2
openai>=1.0.0
tiktoken
sentence-transformers
python-dotenv>=1.0.0
